Combine `usePersistState` into `useLocalStorage`
